requirements.txt
PyQt6
mysql-connector-python
python-dotenv
